Cambio el archivo app.js para dividirlo en rutas (usuarios, comentarios y recetas)
